+++ internal/api/sdconnect_test.go
@@ -16,9 +16,11 @@ type mockConnecter struct {
 	sTokens     map[string]sToken
 	projects    []Metadata
 	projectsErr error
+	token       sToken
+	tokenErr    error
 }
 
-func (c *mockConnecter) getProjects(string, string) ([]Metadata, error) {
+func (c *mockConnecter) getProjects() ([]Metadata, error) {
 	if c.projectsErr != nil {
 		return nil, fmt.Errorf("getProjects error: %w", c.projectsErr)
 	}
@@ -26,7 +28,15 @@ func (c *mockConnecter) getProjects(string, string) ([]Metadata, error) {
 	return c.projects, nil
 }
 
-func (c *mockConnecter) getSTokens([]Metadata, string, string) map[string]sToken {
+func (c *mockConnecter) getToken(string) (sToken, error) {
+	if c.tokenErr != nil {
+		return sToken{}, fmt.Errorf("getToken error: %w", c.tokenErr)
+	}
+
+	return c.token, nil
+}
+
+func (c *mockConnecter) getSTokens([]Metadata) map[string]sToken {
 	return c.sTokens
 }
 
@@ -53,7 +63,7 @@ func Test_SDConnect_GetProjects(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.testname, func(t *testing.T) {
-			MakeRequest = func(url string, query map[string]string, headers map[string]string, body io.Reader, ret any) error {
+			MakeRequest = func(url string, query, headers map[string]string, body io.Reader, ret any) error {
 				if token, ok := headers["X-Authorization"]; !ok || token != "Basic "+tt.token {
 					return fmt.Errorf("Incorrect header 'X-Authorization'\nExpected=%s\nReceived=%s", "Bearer "+tt.token, token)
 				}
@@ -68,8 +78,9 @@ func Test_SDConnect_GetProjects(t *testing.T) {
 				}
 			}
 
-			c := connecter{}
-			projects, err := c.getProjects("https://data.csc.fi", tt.token)
+			url := "https://data.csc.fi"
+			c := connecter{url: &url, token: &tt.token}
+			projects, err := c.getProjects()
 
 			if err != nil {
 				t.Errorf("Unexpected error: %s", err.Error())
@@ -84,12 +95,14 @@ func Test_SDConnect_GetProjects_Error(t *testing.T) {
 	origMakeRequest := MakeRequest
 	defer func() { MakeRequest = origMakeRequest }()
 
-	MakeRequest = func(url string, query map[string]string, headers map[string]string, body io.Reader, ret any) error {
+	MakeRequest = func(url string, query, headers map[string]string, body io.Reader, ret any) error {
 		return errExpected
 	}
 
-	c := connecter{}
-	projects, err := c.getProjects("url", "token")
+	url := "url"
+	token := "token"
+	c := connecter{url: &url, token: &token}
+	projects, err := c.getProjects()
 
 	if err == nil {
 		t.Error("Function should have returned error")
@@ -106,18 +119,19 @@ func Test_SDConnect_GetSTokens(t *testing.T) {
 	var tests = []struct {
 		testname string
 		projects []Metadata
+		override bool
 		sTokens  map[string]sToken
 	}{
 		{
-			"OK_1", []Metadata{{56, "project1"}, {67, "project2"}},
+			"OK_1", []Metadata{{56, "project1"}, {67, "project2"}}, false,
 			map[string]sToken{"project1": {"vhjk", "cud7"}, "project2": {"d6l", "88x6l"}},
 		},
 		{
-			"OK_2", []Metadata{{23, "pr1568"}, {90, "pr2097"}},
+			"OK_2", []Metadata{{23, "pr1568"}, {90, "pr2097"}}, true,
 			map[string]sToken{"pr1568": {"6rxy", "7cli87t"}, "pr2097": {"7cek", "25c8"}},
 		},
 		{
-			"FAIL_STOKENS", []Metadata{{496, "pr152"}, {271, "pr375"}, {12, "pr225"}},
+			"FAIL_STOKENS", []Metadata{{496, "pr152"}, {271, "pr375"}, {12, "pr225"}}, false,
 			map[string]sToken{"pr225": {"8vgicö", "xfd6"}},
 		},
 	}
@@ -127,26 +141,36 @@ func Test_SDConnect_GetSTokens(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.testname, func(t *testing.T) {
-			MakeRequest = func(url string, query map[string]string, headers map[string]string, body io.Reader, ret any) error {
-				if token, ok := headers["X-Authorization"]; !ok || token != "Basic token" {
-					return fmt.Errorf("Real error ccurred")
-				}
+			MakeRequest = func(url string, query, headers map[string]string, body io.Reader, ret any) error {
 				if _, ok := tt.sTokens[query["project"]]; !ok {
 					return fmt.Errorf("Error occurred")
+				}
+				if token, ok := headers["X-Authorization"]; !ok || token != "Basic token" {
+					t.Errorf("X-Authorization header incorrect.")
+				}
+
+				switch project, ok := headers["X-Project-Name"]; {
+				case tt.override && (!ok || project != query["project"]):
+					t.Errorf("X-Project-Name header incorrect.")
+				case !tt.override && ok:
+					t.Errorf("Header should not contain X-Project-Name")
 				}
 
 				switch v := ret.(type) {
 				case *sToken:
 					*v = tt.sTokens[query["project"]]
-
-					return nil
 				default:
-					return fmt.Errorf("ret has incorrect type %v, expected *SToken", reflect.TypeOf(v))
+					t.Errorf("ret has incorrect type %v, expected *SToken", reflect.TypeOf(v))
 				}
+
+				return nil
 			}
 
-			c := connecter{}
-			newSTokens := c.getSTokens(tt.projects, "url", "token")
+			url := "url"
+			token := "token"
+			overridden := tt.override
+			c := connecter{url: &url, token: &token, overriden: &overridden}
+			newSTokens := c.getSTokens(tt.projects)
 
 			if !reflect.DeepEqual(newSTokens, tt.sTokens) {
 				t.Errorf("sTokens incorrect.\nExpected=%s\nReceived=%s", tt.sTokens, newSTokens)
@@ -219,11 +243,17 @@ func Test_SDConnect_GetEnvs(t *testing.T) {
 			err := sd.getEnvs()
 
 			// Test results
-			if err != nil {
-				if err.Error() != tt.expectedError.Error() {
-					t.Errorf("Function returned incorrect error\nExpected=%v\nReceived=%v", tt.expectedError, err)
+			switch {
+			case tt.testname == "OK":
+				if err != nil {
+					t.Errorf("Unexpected error: %s", err.Error())
 				}
+			case err == nil:
+				t.Error("Function did not return error")
+			case err.Error() != tt.expectedError.Error():
+				t.Errorf("Function failed\nExpected=%s\nReceived=%s", tt.expectedError.Error(), err.Error())
 			}
+
 			if sd.url != tt.expectedURL {
 				t.Errorf("URL incorrect. Expected=%v, received=%v", tt.expectedURL, sd.url)
 			}
@@ -236,12 +266,15 @@ func Test_SDConnect_ValidateLogin_OK(t *testing.T) {
 	mockC := &mockConnecter{sTokens: map[string]sToken{"s1": {"sToken", "proj1"}}, projects: projects}
 	sd := &sdConnectInfo{connectable: mockC}
 
-	err := sd.validateLogin("dXNlcjpwYXNz")
+	err := sd.validateLogin("dXNlcjpwYXNz", "")
 	if err != nil {
-		t.Errorf("Function failed, expected no error, received=%v", err)
+		t.Fatalf("Function failed, expected no error, received=%v", err)
 	}
 	if sd.token != "dXNlcjpwYXNz" {
 		t.Errorf("Token incorrect. Expected=dXNlcjpwYXNz, received=%s", sd.token)
+	}
+	if sd.overriden {
+		t.Error("Projects should not have been overridden")
 	}
 	if st := sd.sTokens["s1"].Token; st != "sToken" {
 		t.Errorf("sToken incorrect for project 's1'. Expected=sToken, received=%s", st)
@@ -254,16 +287,71 @@ func Test_SDConnect_ValidateLogin_OK(t *testing.T) {
 	}
 }
 
+func Test_SDConnect_ValidateLogin_Override_OK(t *testing.T) {
+	project := "project_7376"
+	projects := []Metadata{{-1, project}}
+	mockC := &mockConnecter{
+		projectsErr: errors.New("should not have fetched projects"),
+		token:       sToken{"sToken", "projectID"},
+	}
+	sd := &sdConnectInfo{connectable: mockC}
+
+	err := sd.validateLogin("vfylxr7pckgh", project)
+	if err != nil {
+		t.Fatalf("Function failed, expected no error, received=%v", err)
+	}
+	if sd.token != "vfylxr7pckgh" {
+		t.Errorf("Token incorrect. Expected=vfylxr7pckgh, received=%s", sd.token)
+	}
+	if !sd.overriden {
+		t.Error("Projects should have been overridden.")
+	}
+	if st := sd.sTokens[project].Token; st != "sToken" {
+		t.Errorf("sToken incorrect for project '%s'. Expected=sToken, received=%s", project, st)
+	}
+	if pi := sd.sTokens[project].ProjectID; pi != "projectID" {
+		t.Errorf("ProjectID incorrect for project '%s'. expected=projectID, received=%s", project, pi)
+	}
+	if !reflect.DeepEqual(sd.projects, projects) {
+		t.Errorf("Projects incorrect\nExpected=%v\nReceived=%v", projects, sd.projects)
+	}
+}
+
 func Test_SDConnect_ValidateLogin_Fail_GetProjects(t *testing.T) {
 	mockC := &mockConnecter{projectsErr: errors.New("Error occurred")}
 	sd := &sdConnectInfo{connectable: mockC}
 
 	expectedError := "Error occurred for SD Connect: getProjects error: Error occurred"
-	err := sd.validateLogin("user", "pass")
-	if err != nil {
-		if err.Error() != expectedError {
-			t.Errorf("Function failed\nExpected=%v\nReceived=%v", expectedError, err)
-		}
+	err := sd.validateLogin("u7c9Cstlv7", "")
+	if err == nil {
+		t.Error("Function did not return error")
+	} else if err.Error() != expectedError {
+		t.Errorf("Function failed\nExpected=%s\nReceived=%s", expectedError, err.Error())
+	}
+}
+
+func Test_SDConnect_ValidateLogin_Fail_GetToken(t *testing.T) {
+	mockC := &mockConnecter{tokenErr: errors.New("Error occurred")}
+	sd := &sdConnectInfo{connectable: mockC}
+
+	expectedError := "Error occurred for SD Connect: getToken error: Error occurred"
+	err := sd.validateLogin("vhy9pcr7til", "project")
+	if err == nil {
+		t.Error("Function did not return error")
+	} else if err.Error() != expectedError {
+		t.Errorf("Function failed\nExpected=%s\nReceived=%s", expectedError, err.Error())
+	}
+}
+
+func Test_SDConnect_ValidateLogin_Fail_Parameters(t *testing.T) {
+	sd := &sdConnectInfo{}
+	expectedError := "validateLogin() called with too few parameters"
+
+	err := sd.validateLogin("one-param")
+	if err == nil {
+		t.Error("Function did not return error")
+	} else if err.Error() != expectedError {
+		t.Errorf("Function failed\nExpected=%s\nReceived=%s", expectedError, err.Error())
 	}
 }
 
@@ -272,11 +360,11 @@ func Test_SDConnect_ValidateLogin_No_Projects(t *testing.T) {
 	sd := &sdConnectInfo{connectable: mockC}
 
 	expectedError := "No projects found for SD Connect"
-	err := sd.validateLogin("user", "pass")
-	if err != nil {
-		if err.Error() != expectedError {
-			t.Errorf("Function failed\nExpected=%v\nReceived=%v", expectedError, err)
-		}
+	err := sd.validateLogin("f60ovguTit7", "")
+	if err == nil {
+		t.Error("Function did not return error")
+	} else if err.Error() != expectedError {
+		t.Errorf("Function failed\nExpected=%s\nReceived=%s", expectedError, err.Error())
 	}
 }
 
@@ -285,11 +373,11 @@ func Test_SDConnect_ValidateLogin_401_Error(t *testing.T) {
 	sd := &sdConnectInfo{connectable: mockC}
 
 	expectedError := "SD Connect login failed: getProjects error: API responded with status 401 Unauthorized"
-	err := sd.validateLogin("user", "pass")
-	if err != nil {
-		if err.Error() != expectedError {
-			t.Errorf("Function failed\nExpected=%v\nReceived=%v", expectedError, err)
-		}
+	err := sd.validateLogin("69vdtulvf6", "")
+	if err == nil {
+		t.Error("Function did not return error")
+	} else if err.Error() != expectedError {
+		t.Errorf("Function failed\nExpected=%s\nReceived=%s", expectedError, err.Error())
 	}
 }
 
@@ -298,11 +386,11 @@ func Test_SDConnect_ValidateLogin_500_Error(t *testing.T) {
 	sd := &sdConnectInfo{connectable: mockC}
 
 	expectedError := "SD Connect is not available, please contact CSC servicedesk: getProjects error: API responded with status 500 Internal Server Error"
-	err := sd.validateLogin("user", "pass")
-	if err != nil {
-		if err.Error() != expectedError {
-			t.Errorf("Function failed\nExpected=%v\nReceived=%v", expectedError, err)
-		}
+	err := sd.validateLogin("7vr6lvgil", "")
+	if err == nil {
+		t.Error("Function did not return error")
+	} else if err.Error() != expectedError {
+		t.Errorf("Function failed\nExpected=%s\nReceived=%s", expectedError, err.Error())
 	}
 }
 
@@ -599,13 +687,17 @@ func Test_SDConnect_UpdateAttributes_Error(t *testing.T) {
 func Test_SDConnect_DownloadData_Pass(t *testing.T) {
 	// Mock
 	expectedBody := []byte("hellothere")
-	expectedHeaders := map[string]string{"Range": "bytes=0-9", "X-Authorization": "Bearer token", "X-Project-ID": "project"}
+	expectedHeaders := map[string]string{
+		"Range":           "bytes=0-9",
+		"X-Authorization": "Bearer token",
+		"X-Project-ID":    "project",
+	}
 	origMakeRequest := MakeRequest
 	defer func() { MakeRequest = origMakeRequest }()
 	MakeRequest = func(url string, query, headers map[string]string, body io.Reader, ret any) error {
 		// Test that headers were computed properly
 		if !reflect.DeepEqual(headers, expectedHeaders) {
-			t.Errorf("Function failed, expected=%s, received=%s", expectedHeaders, headers)
+			t.Errorf("Function failed\nExpected=%s\nReceived=%s", expectedHeaders, headers)
 		}
 		_, _ = io.ReadFull(bytes.NewReader(expectedBody), ret.([]byte))
 
@@ -621,16 +713,23 @@ func Test_SDConnect_DownloadData_Pass(t *testing.T) {
 		t.Fatalf("Function failed, expected no error, received=%v", err)
 	}
 	if !bytes.Equal(buf, expectedBody) {
-		t.Errorf("Function failed, expected=%s, received=%s", string(expectedBody), string(buf))
+		t.Errorf("Function failed\nExpected=%s\nReceived=%s", string(expectedBody), string(buf))
 	}
 }
 
 func Test_SDConnect_DownloadData_Pass_TokenExpired(t *testing.T) {
 	// Mock
 	expectedBody := []byte("hellothere")
-	expectedHeaders := map[string]string{"Range": "bytes=0-9", "X-Authorization": "Bearer freshToken", "X-Project-ID": "projectID"}
+	expectedHeaders := map[string]string{
+		"Range":           "bytes=0-9",
+		"X-Authorization": "Bearer freshToken",
+		"X-Project-ID":    "projectID",
+		"X-Project-Name":  "project",
+	}
+
 	origMakeRequest := MakeRequest
 	defer func() { MakeRequest = origMakeRequest }()
+
 	MakeRequest = func(url string, query, headers map[string]string, body io.Reader, ret any) error {
 		if token, ok := headers["X-Authorization"]; ok && token == "Bearer freshToken" {
 			// Test that headers were computed properly
@@ -649,6 +748,7 @@ func Test_SDConnect_DownloadData_Pass_TokenExpired(t *testing.T) {
 		connectable: mockC,
 		sTokens:     map[string]sToken{"project": {"expiredToken", "projectID"}},
 		projects:    []Metadata{},
+		overriden:   true,
 	}
 
 	// Test